feat: implement permission-based authorization and refine audits

+++ app/Http/Controllers/Articles/CategoriesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Articles;
 
 use App\Http\Controllers\Controller;
+use App\Http\Controllers\Traits\HasPermissionTrait;
 use App\Models\ArticleCategories;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -10,6 +11,7 @@ use Illuminate\Support\Facades\Validator;
 
 class CategoriesController extends Controller
 {
+  use HasPermissionTrait;
   function getAll()
   {
     $categories = ArticleCategories::orderBy('created_at','DESC')->get();
@@ -70,7 +72,22 @@ class CategoriesController extends Controller
 
   function delete($id)
   {
+    if (!$this->hasPermission('ARTICLE_DELETE_CATEGORY')) {
+      return response()->json([
+        'success' => false,
+        'message' => 'You do not have permission to delete this category.'
+      ], 403);
+    }
+
     $category = ArticleCategories::where('id',$id)->first();
+
+    if (!$category) {
+      return response()->json([
+        'success' => false,
+        'message' => 'Category not found'
+      ], 404);
+    }
+
     $category->delete();
 
     return response()->json('successfully deleted',200);
