+++ app/Http/Controllers/Blogs/BlogCategoriesController.php
@@ -4,12 +4,14 @@ namespace App\Http\Controllers\Blogs;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Http\Controllers\Traits\HasPermissionTrait;
 use App\Models\BlogCategories;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class BlogCategoriesController extends Controller
 {
+    use HasPermissionTrait;
     function getAll()
     {
         $categories = BlogCategories::orderBy('created_at', 'DESC')->get();
@@ -70,7 +72,22 @@ class BlogCategoriesController extends Controller
 
     function delete($id)
     {
+        if (!$this->hasPermission('BLOG_DELETE_CATEGORY')) {
+            return response()->json([
+                'success' => false,
+                'message' => 'You do not have permission to delete this category.'
+            ], 403);
+        }
+
         $category = BlogCategories::where('id', $id)->first();
+
+        if (!$category) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Category not found'
+            ], 404);
+        }
+
         $category->delete();
 
         return response()->json('successfully deleted', 200);
